Added role controller

+++ src/main/java/com/example/service/RoleFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Role;
 
 public interface RoleFetchService {
-	Role fetchById(int id);
+	Role fetchById(Long id);
 
 }
